requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
psycopg2-binary==2.9.9
cloudinary==1.36.0
APScheduler==3.10.4
redis==5.0.1
python-dotenv==1.0.0
Pillow==10.2.0
gunicorn==21.2.0
reportlab==4.2.5